Changes to be committed: 	new file:   additional_attackmodules/homoglyph_v2_attack.py 	new file:   additional_attackmodules/payload_mask_attack.py
Added additional attackmodules for moonshot-data. This commit (and branch) is only meant t
o provide the files as the procedure for contributing attackmodules
is not explicitly stated in the docs.

Homoglyphv2 attackmodule: Modfied orginal homoglyph attack now randomly and sequentially increase the amount of replaced letters with a homoglyph.

Payload Mask: To get the LLM to echo back a prompt that would normally be caught by the prompt filter or embeddings.